ghwawzi/requirements/development.txt
# Development requirements
-r base.txt

# Development tools
Flask-DebugToolbar==0.13.1
watchdog==3.0.0

# Testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
coverage==7.3.2

# Code quality
flake8==6.1.0
black==23.9.1
isort==5.12.0
mypy==1.6.1

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Development utilities
ipython==8.16.1
jupyter==1.0.0